+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/controller/BuyController.java
@@ -1,8 +1,8 @@
 package by.tms.eshopspringboot.controller;
 
-import by.tms.eshopspringboot.model.Order;
-import by.tms.eshopspringboot.model.Product;
-import by.tms.eshopspringboot.model.User;
+import by.tms.eshopspringboot.entity.Order;
+import by.tms.eshopspringboot.entity.Product;
+import by.tms.eshopspringboot.entity.User;
 import by.tms.eshopspringboot.service.OrderServiceAware;
 import by.tms.eshopspringboot.service.ProductServiceAware;
 import lombok.RequiredArgsConstructor;
