+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/controller/CartController.java
@@ -1,6 +1,6 @@
 package by.tms.eshopspringboot.controller;
 
-import by.tms.eshopspringboot.model.Product;
+import by.tms.eshopspringboot.entity.Product;
 import by.tms.eshopspringboot.service.ProductServiceAware;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
