+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/entity/Category.java
@@ -1,4 +1,4 @@
-package by.tms.eshopspringboot.model;
+package by.tms.eshopspringboot.entity;
 
 import java.util.List;
