+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/entity/enums/UserType.java
@@ -1,0 +1,5 @@
+package by.tms.eshopspringboot.entity.enums;
+
+public enum UserType {
+    ADMIN, CLIENT
+}
