--- EshopSpringBoot/src/main/java/by/tms/eshopspringboot/model/enums/UserType.java
@@ -1,5 +1,0 @@
-package by.tms.eshopspringboot.model.enums;
-
-public enum UserType {
-    ADMIN, CLIENT
-}
