+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package by.tms.eshopspringboot.repository;
 
-import by.tms.eshopspringboot.model.Order;
+import by.tms.eshopspringboot.entity.Order;
 
 import java.util.List;
