+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/service/CategoryServiceAware.java
@@ -1,6 +1,6 @@
 package by.tms.eshopspringboot.service;
 
-import by.tms.eshopspringboot.model.Category;
+import by.tms.eshopspringboot.entity.Category;
 
 import java.util.List;
