+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/service/OrderServiceAware.java
@@ -1,6 +1,6 @@
 package by.tms.eshopspringboot.service;
 
-import by.tms.eshopspringboot.model.Order;
+import by.tms.eshopspringboot.entity.Order;
 
 import java.util.List;
