+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/service/UserServiceAware.java
@@ -1,6 +1,6 @@
 package by.tms.eshopspringboot.service;
 
-import by.tms.eshopspringboot.model.User;
+import by.tms.eshopspringboot.entity.User;
 
 public interface UserServiceAware {
     void addUser(User user);
