+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/service/impl/CategoryService.java
@@ -1,6 +1,6 @@
 package by.tms.eshopspringboot.service.impl;
 
-import by.tms.eshopspringboot.model.Category;
+import by.tms.eshopspringboot.entity.Category;
 import by.tms.eshopspringboot.repository.CategoryRepository;
 import by.tms.eshopspringboot.service.CategoryServiceAware;
 import lombok.RequiredArgsConstructor;
